crypto_tracking/requirements.txt
amqp==2.6.0
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
chardet==3.0.4
Django==2.1
djangorestframework==3.11.0
future==0.18.2
idna==2.9
kombu==4.6.10
pytz==2020.1
requests==2.23.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
